10.1
1. Добавлены тесты
2. Изменен тип возращаемого исключения

+++ program/messageRepositories/ustrirovmessagerepository.cpp
@@ -12,7 +12,7 @@ UstrirovMessageRepository::~UstrirovMessageRepository()
 
 }
 
-void UstrirovMessageRepository::SetNormalFvcoRx(quint32 fvcoRx) noexcept
+void UstrirovMessageRepository::SetNormalFvcoRx(double fvcoRx) noexcept
 {
     m_normalMessage.FvcoRx = fvcoRx;
 }
